+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to Rust's notion of
 ## [Unreleased]
 ### Changed
 - `halo2::plonk::Error` has been overhauled:
+  - `Error` now implements `std::fmt::Display` and `std::error::Error`.
   - `Error` no longer implements `PartialEq`. Tests can check for specific error
     cases with `assert!(matches!(..))`.
   - `Error::IncompatibleParams` is now `Error::InvalidInstances`.